Se crea nuevo components menu pero solo la base y se cambio de carpeta formAsk
