+++ src/components/menu.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Menu = () => {
+    return(
+         <h1>aqui va el menu</h1>
+    );
+}
+
+return Menu;
